fixed oauth expires at function to unix time uint64 timestamp (seconds)

+++ include/librespot/oauth.h
@@ -49,7 +49,7 @@ extern OAuth* oauth_build(OAuthBuilder* oauth_builder);
 extern void oauth_free(OAuth* oauth);
 extern const char* oauth_access_token(OAuth* oauth);
 extern const char* oauth_refresh_token(OAuth* oauth);
-extern uint32_t oauth_expires_at(OAuth* oauth);
+extern uint64_t oauth_expires_at(OAuth* oauth);
 
 #ifdef __cplusplus
 }
